Moving Project to use yaml config. Working on #78
