+++ spec/DevBoard/Core/Project/Entity/ProjectRepositorySpec.php
@@ -1,5 +1,5 @@
 <?php
-namespace spec\DevBoard\Core\Project;
+namespace spec\DevBoard\Core\Project\Entity;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Mapping\ClassMetadata;
@@ -10,7 +10,7 @@ class ProjectRepositorySpec extends ObjectBehavior
 {
     public function it_is_initializable()
     {
-        $this->shouldHaveType('DevBoard\Core\Project\ProjectRepository');
+        $this->shouldHaveType('DevBoard\Core\Project\Entity\ProjectRepository');
     }
 
     public function let(EntityManager $em, ClassMetadata $classMetadata)
