+++ src/DevBoard/Behat/Core/Project/FixtureContext.php
@@ -2,7 +2,7 @@
 namespace DevBoard\Behat\Core\Project;
 
 use Behat\Gherkin\Node\TableNode;
-use DevBoard\Core\Project\Project;
+use DevBoard\Core\Project\Entity\Project;
 use Resources\Behat\DefaultContext;
 
 /**
